Update Anime Controller
Took 22 minutes

+++ src/main/java/xyz/themanusia/brot/helper/Helper.java
@@ -17,6 +17,6 @@ public class Helper {
     }
 
     public static String percentage(double v) {
-        return toString(v * 100);
+        return toString((int) (v * 100));
     }
 }
